New version get started

+++ component/Modals/CreateModal/index.js
@@ -5,35 +5,35 @@ import {useRouter} from 'next/router';
 import styles from './styles.module.css';
 
 import Button from '../../Button';
-import { getRoom } from '../../../lib/room';
+import { getRoom } from '../../../lib/roomCreate';
 import { getUser } from '../../../lib/userCreate';
-import { getUserRandom, checkUser } from '../../../lib/anonymousUser';
 
 export default function CreateModal(props) { 
   const [formState, setFormState] = React.useState({
-    roomID: "",
     username: "",
-    password: "",
   });
 
   const Router = useRouter();
 
   const onEnter = async () => {
     localStorage.setItem("username", formState.username);
-    localStorage.setItem("password", formState.password);
-    localStorage.setItem("roomID", formState.roomID);
 
     console.log("Sent: ", formState);
 
-    const user = await getUser(formState);
+    const userExists = await getUser(formState);
 
-    if(!user) var room = await getRoom(formState);
+    var room = null;
+    if(!userExists){
+      roomID = await getRoom(formState);
+      localStorage.setItem("roomID", formState.roomID);
+      console.log("Room Switch: ", room)
+      
+      Router.push({pathname: "/buffer", query: {roomID: roomID, username: formState.username}});
+    }
+    else{
+      console.log("User Already Exists");
+    }
 
-    console.log("Room Switch: ", room)
-    if (room.isBuffer)
-      Router.push({pathname: "/buffer", query: {roomID: formState.roomID, username: formState.username, password: formState.password}});
-    else
-      Router.push({pathname: "/room", query: {roomID: formState.roomID, username: formState.username, password: formState.password}});
   }
 
   return (
@@ -45,22 +45,11 @@ export default function CreateModal(props) {
     >
       <Modal.Header closeButton>
         <Modal.Title id="contained-modal-title-vcenter">
-          Enter Room ID
+          Create a Room
         </Modal.Title>
       </Modal.Header>
       <Modal.Body>
         <Form>
-          <Form.Group controlId="roomID">
-            <Form.Label>Room ID</Form.Label>
-            <Form.Control 
-              type="text" 
-              placeholder="eg: AEF123" 
-              onChange={(e) => 
-                setFormState({ ...formState, roomID: e.target.value})
-              }
-              value={formState.roomID}
-            />
-          </Form.Group>
           <Form.Group controlId="username">
             <Form.Label>Username</Form.Label>
             <Form.Control 
@@ -70,17 +59,6 @@ export default function CreateModal(props) {
                 setFormState({ ...formState, username: e.target.value})
               }
               value={formState.username}
-            />
-          </Form.Group>
-          <Form.Group controlId="password">
-            <Form.Label>Password</Form.Label>
-            <Form.Control 
-              type="text" 
-              placeholder="Set your password for this room" 
-              onChange={(e) => 
-                setFormState({ ...formState, password: e.target.value})
-              }
-              value={formState.password}
             />
           </Form.Group>
         </Form>
